+++ libs/features/anon-apex/eslint.config.js
@@ -1,9 +1,7 @@
 const baseConfig = require('../../../eslint.config.js');
-const nx = require('@nx/eslint-plugin');
 
 module.exports = [
   ...baseConfig,
-  ...nx.configs['flat/react'],
   {
     files: ['**/*.ts', '**/*.tsx', '**/*.js', '**/*.jsx'],
     // Override or add rules here